requirements.txt
scrapy==2.5.0
numpy==1.21.1
pandas==1.3.1
boto3==1.18.16
requests==2.26.0
google-auth-oauthlib==0.4.5
google-api-python-client==2.15.0
google-auth-httplib2==0.1.0
scrapy-inline-requests==0.3.1
w3lib==1.22.0